Disable MyFirebaseMessagingService
MyFirebaseMessagingService works really well for when the app
is already open and the user gets a notification, except
that they'll get the notification even if they're already in
the chat the notification is for. Instead I've switched to
a background notification system and will leave
MyFirebaseMessagingService.kt in the project but deactivated.